refactor: Rename buildinfo package
