+++ main.go
@@ -7,7 +7,7 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/TwiN/gatus/v5/build"
+	"github.com/TwiN/gatus/v5/buildinfo"
 	"github.com/TwiN/gatus/v5/config"
 	"github.com/TwiN/gatus/v5/controller"
 	"github.com/TwiN/gatus/v5/metrics"
@@ -27,7 +27,7 @@ func main() {
 		logr.Infof("Delaying start by %d seconds", delayInSeconds)
 		time.Sleep(time.Duration(delayInSeconds) * time.Second)
 	}
-	buildInfo := build.GetBuildInfo()
+	buildInfo := buildinfo.Get()
 	logr.Infof("Starting Gatus (version: %s, commit: %s, built at: %s)", buildInfo.Version, buildInfo.CommitHash, buildInfo.Time)
 	configureLogging()
 	cfg, err := loadConfiguration()
